refactor: remove student profile components and related functionality
- Deleted StudentProfile, StudentAccordion, ActionPopup, and SuccessMessage components.
- Cleaned up associated imports and state management logic.
- Prepared for a new implementation or restructuring of the student profile feature.
